Fix issue where matrix page is loaded even when graph button is clicked
In the LNT run page, if you clicked the matrix button, then back, then the graph button, you would get the matrix page a second time. This is because we were changing the whole form's submit URL.  Change that to change the submit URL for both graph and matrix, so graph will work after matrix is clicked.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@299135 91177308-0d34-0410-b5e6-96231b3b80d8