Fix LNT’s js to corrrectly determine base URL when server has a prefix
There was a number of errors in the JS leading to a invalid prefix being
generated at runtime. That made some of the URLs and ajax requests
invalid. The prefix was being overwritten, and had the wrong value.

Now we use a global determined by Flask (which should know these things)
and update the urls to not overwrite each other.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@302269 91177308-0d34-0410-b5e6-96231b3b80d8