Add postgres_wrapper.sh to enable tests without server
Add a wrapper script that creates an ad-hoc postgres instance so you can
run the unittests without setting up a real server on your machine.

You still need to have postgres installed and need to enable the test
with `lit -Dpostgres=1`, but at least there is not need to setup and
start a server anymore.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@306326 91177308-0d34-0410-b5e6-96231b3b80d8