Add a delete method with auth to the runs API
Add a token based auth system to the API.  Add a delete method to the
runs endpoint for authorized users.  This will make it easier for non
admins to delete their data.  Hopefully we can get a better auth system
in place, but for now it is single key per instance.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@306901 91177308-0d34-0410-b5e6-96231b3b80d8