Load TestSuite descriptions from json files
The possibility to create custom test suite shemas was already present
in LNT; however it was inconvenient to use as it required the
adminitrator to directly insert descriptions into some (meta) tables in
the database. This adds a more accessible solution:

This creates a new $INSTANCEPATH/schemas directory in which administrators
can place .yaml files describing a testsuite schema. The idea is that
those can be distributed with the testsuites and administrators can copy
or symlink the schema into their server instances.

See the added example in the documentation for an example of the format.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@307222 91177308-0d34-0410-b5e6-96231b3b80d8