Schema migration logic for custom schemas
For yaml based testsuite schemas: This saves the last schema used in the
database. When the schema in the database differs from the configured
schema a migration strategy is tried: New metrics, run_fields and
machine_field columns are added as needed. If there are any other
differences in the schema the schema is rejected and server startup
aborted.

This also changes the default value for run fields and machine fields
from '' to null/None to better match the effects when adding a new
column. The rest of the UI seems unaffected by this.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@307223 91177308-0d34-0410-b5e6-96231b3b80d8