api: Add machine merge operation
Add api operation to merge 1 machine into another.

Duplicated machines are a very common occurence with the current
behaviour where we create a new one when parameters don't match, so it
is very helpful to have a way to clean this up by merging two machines
(typically with the same name).

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@308709 91177308-0d34-0410-b5e6-96231b3b80d8