Fix for MySQL case insensitive queries
MySQL Collation is by default  'utf8_general_ci' which is not case
sensitive. In sample creation, we were making a dict of the test names,
and using this to check if new Test were added. Because the DB compare
is case sensitive, and the Python string compare is case sensitive,
when we had a test were the case of the test name changed MySQL will
issue an integrity error.  To fix, hav eht tables use the utf8_bin
collation which is case sensitive.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@312400 91177308-0d34-0410-b5e6-96231b3b80d8