Change duplicate submission default to 'reject'
The previous default of 'replace' would delete the previous run in case
of a duplicate submission (a submission to the same machine with the
same order number). This caused confusion to unsuspecting users as runs
would magically disappear. Change the default to 'reject' so it is more
obvious that something was strange on the submission side.

git-svn-id: https://llvm.org/svn/llvm-project/lnt/trunk@317389 91177308-0d34-0410-b5e6-96231b3b80d8